backend/requirements.txt
flask
flask-cors
bcrypt
requests
python-dotenv
astor
